pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["AlifahadRajpoot <Ranaalifahad.fsd.275@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.111.0"
uvicorn = {extras = ["standards"], version = "^0.29.0"}
sqlmodel = "^0.0.18"
python-dotenv = "^1.0.1"
sqlalchemy = "^2.0.30"
psycopg2-binary = "^2.9.9"

[tool.poetry.scripts]
dev="app.main:start"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
